# Attempting to use unsafe.Slice again