In two commits:

 - 80d6073819895a4ee0502c0f21996f16a5afb390
 - 9005c391b9cbdb19f2e83362c336f51a4bdc9807

All usages of unsafe.Slice were removed due to performance reason,
without any proof. In any case, we should prefer correctness over
performance, code should be done right, before thinking about speed.

However, this commit attempts to use unsafe.Slice again, and adding
benchmarks. The result indicates that unsafe.Slice even slightly faster
than incorrect usage of reflect.SliceHeader:

```
name                     old time/op    new time/op    delta
HaxMapReadsOnly-8          10.6µs ± 1%    10.2µs ± 1%  -3.72%  (p=0.000 n=9+10)
HaxMapReadsWithWrites-8    12.6µs ± 2%    12.1µs ± 1%  -4.18%  (p=0.000 n=10+9)

name                     old alloc/op   new alloc/op   delta
HaxMapReadsOnly-8           0.00B          0.00B         ~     (all equal)
HaxMapReadsWithWrites-8    1.04kB ± 3%    1.01kB ± 4%  -2.76%  (p=0.003 n=10+10)

name                     old allocs/op  new allocs/op  delta
HaxMapReadsOnly-8            0.00           0.00         ~     (all equal)
HaxMapReadsWithWrites-8       130 ± 4%       126 ± 4%  -2.93%  (p=0.004 n=10+10)
```

Note that the benchmark is run with perflock to prevent noisy as much as
possible:

```
perflock -governor=70% go1.19 test -run=NONE -bench=. -benchmem -count=10 .
```