@alphadose 2022-08-29
@cuonglm can you paste the [neofetch](https://github.com/dylanaraps/neofetch) of your system ?