@cuonglm 2022-08-29
@alphadose I don't think the system affects the benchmark, just see how my benchmark is more stable that yours, read-write benchmark just have `± 1-2 %`.