@alphadose 2022-08-29
> @alphadose I don't think the system affects the benchmark, just see how my benchmark is more stable that yours, read-write benchmark just have ± 1-2 %.

I disagree, golang is much more mature on amd64, my system arm64 is something which came up recently hence there is a certain degree of instability.

Although the golang developers are continuously improving the performance on arm64 as seen in their latest patch notes, it seems that for my system `unsafe.Slice` is slower.