@alphadose 2022-08-29
there are lots of unsafe tricks such as this one from golang core https://cs.opensource.google/go/go/+/refs/tags/go1.19:src/math/unsafe.go;l=18

The linter says not using `unsafe.Slice` might lead to a **possible** misuse

but for us there is no problem because everything is allocated and deallocated on the function stack itself, nothing escapes to heap, hence we are not even challenging correctness in the first place.

Not using `unsafe.Slice` might lead to a possible misuse but I don't think it would make a difference in terms of correctness unless there is a bug with proper steps to reproduce the results