@cuonglm 2022-08-29
> it seems that for my system unsafe.Slice is slower.

Note that like I said in the commit, your benchmark does not mean unsafe.Slice is slow, there're too much noise in your benchmark, which affect the benchmark result. Please lock down your CPU, only the result with less than 5% noisy would make sense.