@alphadose 2022-08-31
@cuonglm the latest commit https://github.com/alphadose/haxmap/commit/eee8abe05d83c726178d921bdb498baf74a6e671 fixes all the issues

and it has a lot of performance improvement as well just by using arrays instead of slices 🥳