@alphadose 2022-08-31
> Cool, but any proof? I still see your benchmark result have too much noise.

I cannot draw up cleaner benchmarks in which case you should also run the comparisons in your own environment for an additional perspective.

But theoretically the reason for being fast is golang slices have 2 pointer hops whereas arrays have 1 pointer hop for access hence it is faster.
My current benchmarks although noisy seem to demonstrate this to a certain extent.